requirements.txt
protobuf==6.32.1
psycopg==3.2.10
psycopg_pool==3.2.6
python-dotenv==1.1.1
Requests==2.32.5
google-genai==1.32.0